Add Glowing Object Prefab & set one up in LevelOne
This is a first draft of how to handle picking up objects.

Steps to do this (as best I can remember):

Make an empty game object that has children:
- point light
- sphere, with its collider removed
Add sphere collider to the parent "Glowing Object"
Add "HoldLocation" child to the Player
Add "Clickable" tag to Glowing Object (not yet used, but will be)
Add "PickMeUp" script to Glowing Object
Drag Glowing Object to the Prefabs folder - now it can be reused!