Animate light orbs
The light orbs randomly float around slowly and stop when you pick them up.

It was necessary to add a Rigidbody so the orbs don't pass through the cave walls.

Here I've introduced a bug where now you can collide with the light orb when you're holding it, messing with its position.  I think we need to disable the Rigidbody the way we disable the Collider when pickedUp is true, but Rigidbody doesn't have an "enabled" option . . .