Orbs can no longer get too close to walls!
- Move the physics collider to GlowingObject
This allows us to give it a radius (0.15), so it won't approach walls anymore.  The physics collider has to be on the top-level parent to work properly.
- Put the GlowingObject in the OrbCollider layer
Otherwise it would bounce off the player.  Whereas all children are in the Default layer now.
- Give one of its children—the Point Light—the trigger collider.
The trigger collider needs to be in a layer that can collide with the player (so, a child in the Default layer—rather than GlowingObject itself).  If we used the Sphere, the trigger collider's radius would get multiplied by the Sphere's radius (0.0625), which is totally unrelated.  By placing the trigger collider on the Point Light instead of the Sphere we are able to specify it directly (1.5).
- Modify the PickMeUp and PickUpObject scripts accordingly.
PickUpObject now checks whether the object or its parent has the PickMeUp script.