Blend instead of increasing polygons
Wow, this change made the FPS shoot up!  I discovered that the surfaces can be blended by turning on "calculate normals" instead of "import normals".  This makes the Catmull subdivision unnecessary, which was hurting performance anyway.  I also love the low-poly look on outlines of objects while surfaces look smooth.  I turned on smoothing for the Spelunker as well.

Also turned on normal map surface for the Cave Grey material since performance is better now.

Also removed an inner face of Spelunker that was interfering with smoothing.