Stop sliding when not moving!
By switching to a friction PhysicMaterial.

Using friction all the time would result in the player getting stuck on small slopes.

This change includes a tweak to height, so the player's mesh extends a little below the collider, to fix a gap between the player and the ground on slopes.  Previously, the player's sliding on slopes was a more distracting issue. — This would be better solved with a different collider, but a skinner CapsuleCollider also results in the player getting stuck on the ground sometimes.